Binary CrossEntropy is a wrong choice + Small changes
Important Ones:

- When we use binary_crossentropy, it does not evaluate the log_loss
function, it was evaluating something else, we should realize this
after submissions actually, when the program says loss is around 0.4 we
get around 0.6 from Kaggle. I realized the formula it was trying to
minimize, it is the mean of binary_crossentropy at each output (3 of
them) but log_loss evaluates something different, so the system was not
reporting correct loss and was not minimizing correct loss, I guess it
was harsher optimization, anyway at least with this
categorical_crossentropy when the program outputs 0.6 then Kaggle also
says similar result.

- This one is IMPORTANT: normalization procedure is enhanced   to
represent data more

- This one is also IMPORTANT: We were not normalizing some numeric val
columns, now we are normalizing all the columns

RESULTs improved to ~0.600

- added 1more layer (I don’t want 1024 at the last layer due to
regularization penalty)

- Increased Dropout to 0.5

- Instead of giving activation as parameter (it adds automatically with
default values) I added manually each activation function layer to make
it more flexible for trials

- I changed the optimizer to adadelta, I saw that it has some
advantages compared to adam

- I added more information to output file names